fix: lazy-load exact duplicate resolution
